third_party/WebKit/LayoutTests/bluetooth/descriptor/readValue/descriptor-is-blocklisted.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/testdriver.js"></script>
<script src="../../../resources/testdriver-vendor.js"></script>
<script src="../../../external/wpt/bluetooth/resources/bluetooth-helpers.js"></script>
<script>
'use strict';
const test_desc = 'Attempt to call readValue on a blocked descriptor must ' +
    'generate a SecurityError';
const expected = new DOMException(
    'readValue() called on blocklisted object marked exclude-reads. ' +
    'https://goo.gl/4NeimX',
    'SecurityError');

bluetooth_test(() => getBlocklistExcludeReadsDescriptor()
    .then(({descriptor}) => assert_promise_rejects_with_message(
        descriptor.readValue(),
        expected)),
    test_desc);
</script>
